# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Nexus Audio Player API
  version: 1.0.0
  description: |
    API documentation for Nexus Audio Player - A modern audio/video player with cloud sync capabilities.
    
    ## Authentication
    All protected endpoints require authentication via Bearer token in the Authorization header.
    The token can be either:
    - Firebase ID token (recommended)
    - Google OAuth access token
    
    ## Rate Limiting
    API requests are rate-limited. Contact support if you need higher limits.
    
    ## Error Responses
    All errors follow a consistent format:
    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable error message",
        "details": {}
      }
    }
    ```
  contact:
    name: Nexus Audio Player Support
    email: support@nexus-audio.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.nexus-audio.com/api
    description: Production server
...
paths:
...
  /storage/upload-bunny:
    post:
      tags:
        - Storage
      summary: Upload a file to Bunny
      description: |
        Upload a file to Bunny Storage (Pro only).

        **File size limit:** 500MB
      operationId: uploadFileBunny
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /storage/upload-planethoster:
    post:
      tags:
        - Storage
      summary: Upload a file to PlanetHoster
      description: |
        Upload a file to PlanetHoster SFTP (Pro only).

        **File size limit:** 500MB
      operationId: uploadFilePlanetHoster
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token authentication. Use either:
        - Firebase ID token (recommended)
        - Google OAuth access token
  schemas:
    UploadResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique file ID
          example: "1234567890-123456789"
        url:
          type: string
          format: uri
          description: URL to access the file
          example: "https://cdn.example.com/nexus/user123/file.mp3"
        size:
          type: integer
          description: File size in bytes
          example: 5242880
        filename:
          type: string
          description: Original filename
          example: "song.mp3"
        provider:
          type: string
          enum: [local, bunny]
          description: Storage provider used
          example: "bunny"
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "File too large. Maximum size is 100MB"
            details:
              type: object
              description: Additional error details
              nullable: true
  responses:
    BadRequest:
      description: Bad request - validation error or invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "File too large. Maximum size is 100MB"
              details:
                maxSize: 104857600
                fileSize: 157286400
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "AUTHENTICATION_ERROR"
              message: "User not authenticated"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"
...